最近情绪低落，长时间未提交更新： 1，添加多语言初始化（部分表还不能正常使用，没有统一变更AD_LANGUAGE->AD_LANGUAGE_ID）及xml导入功能 2，将原来使用JPA实体继承方式（会出错）读取多语言数据的方式改为单独使用view进行查询读取

+++ myerpGWT/src/main/java/org/adempiere/model/ADPackageExpDetail.java
@@ -35,7 +35,7 @@ public class ADPackageExpDetail extends org.adempiere.common.ADEntityBase {
 	private Integer aDWorkflowID;
 	private Boolean active;
 	private String created;
-	private BigDecimal createdBy;
+	private Integer createdBy;
 	private String dBType;
 	private String description;
 	private String destinationDirectory;
@@ -304,11 +304,11 @@ public class ADPackageExpDetail extends org.adempiere.common.ADEntityBase {
 
 	@Basic
 	@Column(nullable=false)
-	public BigDecimal getCreatedBy() {
+	public Integer getCreatedBy() {
 		return createdBy;
 	}
 
-	public void setCreatedBy(BigDecimal createdBy) {
+	public void setCreatedBy(Integer createdBy) {
 		this.createdBy = createdBy;
 	}
